+++ src/test/java/tests/browse_languages/BrowseLanguagesSubmenuTest.java
@@ -1,18 +1,13 @@
 package tests.browse_languages;
 
 import base.BaseTest;
-import org.openqa.selenium.WebElement;
 import org.testng.Assert;
-import org.testng.Reporter;
 import org.testng.annotations.Test;
 import pages.browse_languages.BrowseLanguagesSubmenuPage;
 import pages.browse_languages.letters.ABCPage;
 import pages.browse_languages.letters.NPage;
 import utils.StaticProvider;
 
-import java.io.IOException;
-import java.net.HttpURLConnection;
-import java.net.URL;
 import java.util.List;
 
 public class BrowseLanguagesSubmenuTest extends BaseTest {
@@ -62,50 +57,6 @@ public class BrowseLanguagesSubmenuTest extends BaseTest {
         List<String> actualSubmenusNames = browseLanguagesSubmenuPage.getSubmenusNames();
 
         Assert.assertEquals(actualSubmenusNames, expectedlettersSubmenu);
-    }
-
-    @Test
-    public void test_API_AllLanguagesLinksAreNotBroken() {
-        String linkURL = "";
-        int responseCode;
-        int actualWorkingLinksCount = 0;
-
-        List<WebElement> aTags = openBaseURL()
-                .clickBrowseLanguagesMenu()
-                .getSubmenus();
-
-        final int expectedWorkingLinksCount = aTags.size();
-        int internalLinks = expectedWorkingLinksCount;
-
-        for (WebElement link : aTags) {
-            linkURL = link.getAttribute("href");
-
-            if (linkURL != null && !linkURL.isBlank() && !linkURL.isEmpty()) {
-                if (!linkURL.startsWith(getBaseUrl())) {
-                    Reporter.log(linkURL + " is externalLink ", true);
-                    internalLinks--;
-                } else {
-                    try {
-                        HttpURLConnection connection = (HttpURLConnection) (new URL(linkURL).openConnection());
-                        connection.setRequestMethod("HEAD");
-                        connection.connect();
-
-                        responseCode = connection.getResponseCode();
-
-                        if (responseCode < 400) {
-                            actualWorkingLinksCount++;
-                        } else {
-                            Reporter.log(linkURL + " is broken, responseCode " + responseCode, true);
-                        }
-                    } catch (IOException e) {
-                        e.printStackTrace();
-                    }
-                }
-            }
-        }
-
-        Assert.assertEquals(actualWorkingLinksCount, internalLinks);
-        Assert.assertEquals(actualWorkingLinksCount, expectedWorkingLinksCount);
     }
 
     @Test(dataProviderClass = StaticProvider.class, dataProvider = "lettersSubmenu")
